add transaction-user

+++ src/main/java/vn/compedia/website/repository/TransactionRepositoryCustom.java
@@ -13,7 +13,7 @@ public interface TransactionRepositoryCustom {
 
     List<TransactionDto> getAllByUserName(String userName,TransactionSearchDto dto);
 
-    int countSearchByUserName(String userName,TransactionSearchDto searchDto);
+    BigInteger countSearchByUserName(String userName,TransactionSearchDto searchDto);
 
     BigInteger countSearch(TransactionSearchDto searchDto);
 }
